-

UnsteadyFlowSolvers.lautat — Function.

lautat(surf, curfield, nsteps=500, dtstar=0.015, startflag=0,

writeflag=0, writeInterval=1000., delvort=delNone(), maxwrite=100, nround=6)

Simulates potential flow for an airfoil undergoing unsteady motion using Large-Angle Unsteady Thin-Airfoil Theory.

Inputs required:
- `TwoDSurf` : based on airfoil geometry and kinematics
- `TwoDFlowField` : initial flowfield

`dtstar` should be in accordance with the kinematics; see `?find_tstep`.

`nsteps` should be calculated according to total run time required and
dtstar.

`startflag=1` can be used to continue a simulation using the
*resultsSummary* file in current directory.

`writeflag=1` along with a suitable writeInterval is used to write flow
data at intervals. These can be used to create plots and movies of the
flowfield.

`delvort` is used to control the vortex count. The simulation cost
increases as O(n``^2``) with number of vortices in the flow field and so
long run times can get very computationally expensive. Look at
`?delVortDef` for algorthims to control vortex count.

`maxwrite` is the maximum number of timestamps/directories that may be
written.

`nround` is the number of digits to which the output timestamps are
rounded off to.

The Large-Angle Unsteady Thin-Airfoil theory is described in: Ramesh,
K. et al., "An unsteady airfoil theory applied to pitching motions
validated against experiment and computation", Theor. Comput. Fluid
Dyn. (2013) 27: 843. [Weblink](https://doi.org/10.1007/s00162-012-0292-8)

source